+++ e2e/pages/citizen/createCase/C100/otherProceedings/childAbductionOrderDetailsPage.ts
@@ -4,7 +4,7 @@ import { Page } from "@playwright/test";
 import { Helpers } from "../../../../../common/helpers";
 import { CommonStaticText } from "../../../../../common/commonStaticText";
 import { ChildAbductionOrderDetailsContent } from "../../../../../fixtures/citizen/createCase/C100/otherProceedings/childAbductionOrderDetailsContent";
-import { CitizensOtherProceedingsUniqueSelectors } from "../../../../../common/enums";
+import { CitizensOtherProceedingsUniqueSelectors } from "../../../../../common/commonIds";
 
 interface ChildAbductionOrderDetailsPageOptions {
   page: Page;
