+++ e2e/tests/citizen/createCase/C100/c100CitizenApplicationWrittenAgreementYes.spec.ts
